+++ src/converse-headline.js
@@ -8,7 +8,7 @@ import "converse-chatview";
 import converse from "@converse/headless/converse-core";
 import tpl_chatbox from "templates/chatbox.html";
 
-const { _, utils } = converse.env;
+const { _, moment, utils } = converse.env;
 
 
 converse.plugins.add('converse-headline', {
@@ -52,13 +52,21 @@ converse.plugins.add('converse-headline', {
             { __ } = _converse;
 
         _converse.HeadlinesBox = _converse.ChatBox.extend({
-            defaults: {
-                'type': _converse.HEADLINES_TYPE,
-                'bookmarked': false,
-                'chat_state': undefined,
-                'num_unread': 0,
-                'url': ''
+            defaults () {
+                return {
+                    'bookmarked': false,
+                    'hidden': _.includes(['mobile', 'fullscreen'], _converse.view_mode),
+                    'message_type': 'headline',
+                    'num_unread': 0,
+                    'time_opened': this.get('time_opened') || moment().valueOf(),
+                    'type': _converse.HEADLINES_TYPE
+                }
             },
+
+            initialize () {
+                this.initMessages();
+                this.set({'box_id': `box-${btoa(this.get('jid'))}`});
+            }
         });
 
 
@@ -110,7 +118,7 @@ converse.plugins.add('converse-headline', {
             if (utils.isHeadlineMessage(_converse, message)) {
                 const from_jid = message.getAttribute('from');
                 if (_.includes(from_jid, '@') && 
-                        !_converse.api.contacts.get(from_jid) &&
+                        !_converse.roster.get(from_jid) &&
                         !_converse.allow_non_roster_messaging) {
                     return;
                 }
@@ -142,10 +150,10 @@ converse.plugins.add('converse-headline', {
 
 
         _converse.api.listen.on('chatBoxViewsInitialized', () => {
-            const that = _converse.chatboxviews;
+            const views = _converse.chatboxviews;
             _converse.chatboxes.on('add', item => {
-                if (!that.get(item.get('id')) && item.get('type') === _converse.HEADLINES_TYPE) {
-                    that.add(item.get('id'), new _converse.HeadlinesBoxView({model: item}));
+                if (!views.get(item.get('id')) && item.get('type') === _converse.HEADLINES_TYPE) {
+                    views.add(item.get('id'), new _converse.HeadlinesBoxView({model: item}));
                 }
             });
         });